requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
django-rest-framework
pytz==2022.1
sqlparse==0.4.2
pillow
django-crispy-forms
django-admin-interface


